(A101032-30) Fix root logger no longer working.
