+++ src/utils/graph.py
@@ -1,43 +1,8 @@
-import os
-from _tkinter import TclError
-from multiprocessing import Process
-
-import numpy as np
-import matplotlib
 import matplotlib.pyplot as plt
 
 from utils.file import File
 from utils.log import Log
-
-BACKEND_TIMEOUT = 5
-
-
-# TODO(M-Whitaker): This function is in a rubbish place however you can't pre-declare functions in python.
-def check_backend():
-    """
-    If fails will return False and change the backend to headless.
-    :return:
-    """
-    tk = True
-    x = np.arange(0, 5, 0.1)
-    y = np.sin(x)
-    p1 = Process(target=plt.plot, args=(x, y))
-    try:
-        p1.start()
-        p1.join(timeout=BACKEND_TIMEOUT)
-        if p1.exitcode is None:
-            raise TclError("pyplot timed out: could not find DISPLAY \"%s\"" % os.environ['DISPLAY'])
-        matplotlib.use("TKAgg", force=True)
-        tk = True
-    except TclError as e:
-        Log.warning(e)
-        matplotlib.use("Agg", force=True)
-        tk = False
-    finally:
-        p1.kill()
-        plt.close()
-        Log.info("Set backend to %s" % matplotlib.get_backend())
-        return tk
+from utils._graph import check_backend
 
 
 try:
